Adds p-values to the output comparing values across groups

Adds p-values to the output comparing values across groups

add_comparison(x, test = NULL, pvalue_fun = style_pvalue,
  group = x$inputs$group)

Arguments

x
object with class tbl_summary from the tbl_summary function

test
user defined list of statistical tests provided as a named character vector with variables as names and test functions as values., e.g. list(age = "t.test", ptstage = "fisher.test"). Options include "t.test" for a T-test, "wilcox.test" for a Wilcoxon rank sum test, "kruskal.test" for a Kruskal-Wallis rank sum test, "chisq.test" for a Chi-squared test, "fisher.test" for a Fisher's exact test, and "re" for a random intercept model to account for clustered data. For "re" to be used "group" must also be specified in the function call.

pvalue_fun
function for rounding/formatting p-values. Default is style_pvalue. The function must have a single input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, partial(style_pvalue, digits = 2)).

group
Character vector of an ID or grouping variable. Summary statistics will not be printed for this column, but they may be used in subsequent functions. For example, the group column may be used in add_comparison() to include p-values with correlated data. Default is the group = input from tbl_summary

Examples

comp <- trial %>% tbl_summary(by = "trt") %>% add_comparison()