Adds the global p-value for a categorical variables in tbl_uvregression objects

This function uses Anova from the car package with type = "III" to calculate global p-values. If a needed class of model is not supported by Anova, please put in an issue at https://github.com/ddsjoberg/gtsummary/issues to request support.

# S3 method for tbl_uvregression
add_global(x, ...)

Arguments

x
object with class tbl_uvregression from the tbl_uvregression function

...
arguments to be passed to Anova. Adding test.statistic = can change the type of test (e.g. Likelihood-ratio, Wald, etc.).

Examples

tbl_uvregression(
  trial,
  method = "glm",
  y = "response",
  method.args = list(family = binomial),
  exponentiate = TRUE
) %>%
  add_global()

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #hrtfztbnsr .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #hrtfztbnsr .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #hrtfztbnsr .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #hrtfztbnsr .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #hrtfztbnsr .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #hrtfztbnsr .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #hrtfztbnsr .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #hrtfztbnsr .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #hrtfztbnsr .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #hrtfztbnsr .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #hrtfztbnsr .gt_striped {
#>   background-color: #f2f2f2;
#> }
#> 
#> #hrtfztbnsr .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #hrtfztbnsr .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #hrtfztbnsr .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #hrtfztbnsr .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #hrtfztbnsr .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #hrtfztbnsr .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #hrtfztbnsr .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #hrtfztbnsr .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #hrtfztbnsr .gt_center {
#>   text-align: center;
#> }
#> 
#> #hrtfztbnsr .gt_left {
#>   text-align: left;
#> }
#> 
#> #hrtfztbnsr .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #hrtfztbnsr .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #hrtfztbnsr .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #hrtfztbnsr .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #hrtfztbnsr .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #hrtfztbnsr .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="hrtfztbnsr" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>N</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>OR</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>95% CI</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>p-value</strong></th>
#> </tr>
#> <tbody class='gt_table_body'>
#> <tr>
#> <td class='gt_row gt_left'>Treatment Randomization</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.011</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">Drug</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>—</td>
#> <td class='gt_row gt_center gt_striped'>—</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Placebo</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.47</td>
#> <td class='gt_row gt_center'>0.26, 0.84</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>Age, yrs</td>
#> <td class='gt_row gt_center gt_striped'>183</td>
#> <td class='gt_row gt_center gt_striped'>1.00</td>
#> <td class='gt_row gt_center gt_striped'>0.98, 1.0</td>
#> <td class='gt_row gt_center gt_striped'>>0.9</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Marker Level, ng/mL</td>
#> <td class='gt_row gt_center'>183</td>
#> <td class='gt_row gt_center'>0.92</td>
#> <td class='gt_row gt_center'>0.65, 1.3</td>
#> <td class='gt_row gt_center'>0.6</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>T Stage</td>
#> <td class='gt_row gt_center gt_striped'>191</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>0.055</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T1</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">T2</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>1.4</td>
#> <td class='gt_row gt_center gt_striped'>0.63, 3.1</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T3</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.63</td>
#> <td class='gt_row gt_center'>0.26, 1.5</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">T4</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>0.50</td>
#> <td class='gt_row gt_center gt_striped'>0.22, 1.1</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Grade</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.4</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">I</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>—</td>
#> <td class='gt_row gt_center gt_striped'>—</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">II</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.60</td>
#> <td class='gt_row gt_center'>0.29, 1.3</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">III</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>0.76</td>
#> <td class='gt_row gt_center gt_striped'>0.38, 1.5</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> </div>