add_p_ {gtsummary}	R Documentation

Standard evaluation version of add_p()

Description

The 'group =' argument can be passed as a string, rather than with non-standard evaluation as in add_p. Review the help file for add_p fully documented options and arguments.

Usage

add_p_(
  x,
  test = NULL,
  pvalue_fun = NULL,
  group = NULL,
  include = NULL,
  exclude = NULL
)

Arguments

x
Object with class tbl_summary from the tbl_summary function

test
List of formulas specifying statistical tests to perform, e.g. list(all_continuous() ~ "t.test", all_categorical() ~ "fisher.test"). Options include

"t.test" for a t-test,

"aov" for a one-way ANOVA test,

"wilcox.test" for a Wilcoxon rank-sum test,

"kruskal.test" for a Kruskal-Wallis rank-sum test,

"chisq.test" for a chi-squared test of independence,

"chisq.test.no.correct" for a chi-squared test of independence without continuity correction,

"fisher.test" for a Fisher's exact test,

"lme4" for a random intercept logistic regression model to account for clustered data, lme4::glmer(by ~ variable + (1 | group), family = binomial). The by argument must be binary for this option.

Tests default to "kruskal.test" for continuous variables, "chisq.test" for categorical variables with all expected cell counts >=5, and "fisher.test" for categorical variables with any expected cell count <5. A custom test function can be added for all or some variables. See below for an example.

pvalue_fun
Function to round and format p-values. Default is style_pvalue. The function must have a numeric vector input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, purrr::partial(style_pvalue, digits = 2)).

group
Column name of an ID or grouping variable. The column can be used calculate p-values with correlated data (e.g. when the test argument is "lme4"). Default is NULL. If specified, the row associated with this variable is omitted from the summary table.

include
Character vector of variable names to include from output.

exclude
Character vector of variable names to exclude from output.

[Package gtsummary version 1.2.4 Index]