Adds a column showing a label for the summary statistics shown in each row

Rather than simply printing the summary statistics, with the use of add_stat_label(), a column labeling the summary statistics is added.

add_stat_label(x, iqr = TRUE)

Arguments

x
object with class tbl_summary from the tbl_summary function

iqr
logical indicator whether 'p25, p75' should resolve to 'IQR'. Default is TRUE

Examples

mtcars %>% tbl_summary() %>% add_stat_label()

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #ljamkajowk .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #ljamkajowk .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #ljamkajowk .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ljamkajowk .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ljamkajowk .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #ljamkajowk .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #ljamkajowk .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #ljamkajowk .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #ljamkajowk .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #ljamkajowk .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #ljamkajowk .gt_striped {
#>   background-color: #f2f2f2;
#> }
#> 
#> #ljamkajowk .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #ljamkajowk .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #ljamkajowk .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #ljamkajowk .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #ljamkajowk .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #ljamkajowk .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #ljamkajowk .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #ljamkajowk .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #ljamkajowk .gt_center {
#>   text-align: center;
#> }
#> 
#> #ljamkajowk .gt_left {
#>   text-align: left;
#> }
#> 
#> #ljamkajowk .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ljamkajowk .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ljamkajowk .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ljamkajowk .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ljamkajowk .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ljamkajowk .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="ljamkajowk" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Statistic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'>stat_0</th>
#> </tr>
#> <tbody class='gt_table_body'>
#> <tr>
#> <td class='gt_row gt_left'>mpg</td>
#> <td class='gt_row gt_center'>Median (IQR)</td>
#> <td class='gt_row gt_center'>19.2 (15.4, 22.8)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>cyl</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">4</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>11 (34%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">6</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>7 (22%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">8</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>14 (44%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>disp</td>
#> <td class='gt_row gt_center gt_striped'>Median (IQR)</td>
#> <td class='gt_row gt_center gt_striped'>196 (121, 326)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>hp</td>
#> <td class='gt_row gt_center'>Median (IQR)</td>
#> <td class='gt_row gt_center'>123 (96, 180)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>drat</td>
#> <td class='gt_row gt_center gt_striped'>Median (IQR)</td>
#> <td class='gt_row gt_center gt_striped'>3.70 (3.08, 3.92)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>wt</td>
#> <td class='gt_row gt_center'>Median (IQR)</td>
#> <td class='gt_row gt_center'>3.33 (2.58, 3.61)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>qsec</td>
#> <td class='gt_row gt_center gt_striped'>Median (IQR)</td>
#> <td class='gt_row gt_center gt_striped'>17.71 (16.89, 18.90)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>vs</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>14 (44%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>am</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>13 (41%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>gear</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">3</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>15 (47%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">4</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>12 (38%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">5</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>5 (16%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>carb</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">1</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>7 (22%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">2</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>10 (31%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">3</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>3 (9.4%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">4</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>10 (31%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">6</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>1 (3.1%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">8</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>1 (3.1%)</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> </div>

mtcars %>% tbl_summary(by = "am") %>% add_stat_label(iqr = FALSE)

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #npghmxkfrb .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #npghmxkfrb .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #npghmxkfrb .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #npghmxkfrb .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #npghmxkfrb .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #npghmxkfrb .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #npghmxkfrb .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #npghmxkfrb .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #npghmxkfrb .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #npghmxkfrb .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #npghmxkfrb .gt_striped {
#>   background-color: #f2f2f2;
#> }
#> 
#> #npghmxkfrb .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #npghmxkfrb .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #npghmxkfrb .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #npghmxkfrb .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #npghmxkfrb .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #npghmxkfrb .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #npghmxkfrb .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #npghmxkfrb .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #npghmxkfrb .gt_center {
#>   text-align: center;
#> }
#> 
#> #npghmxkfrb .gt_left {
#>   text-align: left;
#> }
#> 
#> #npghmxkfrb .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #npghmxkfrb .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #npghmxkfrb .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #npghmxkfrb .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #npghmxkfrb .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #npghmxkfrb .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="npghmxkfrb" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Statistic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'>stat_1</th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'>stat_2</th>
#> </tr>
#> <tbody class='gt_table_body'>
#> <tr>
#> <td class='gt_row gt_left'>mpg</td>
#> <td class='gt_row gt_center'>Median (25%, 75%)</td>
#> <td class='gt_row gt_center'>17.3 (14.9, 19.2)</td>
#> <td class='gt_row gt_center'>22.8 (21.0, 30.4)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>cyl</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">4</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>3 (16%)</td>
#> <td class='gt_row gt_center'>8 (62%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">6</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>4 (21%)</td>
#> <td class='gt_row gt_center gt_striped'>3 (23%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">8</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>12 (63%)</td>
#> <td class='gt_row gt_center'>2 (15%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>disp</td>
#> <td class='gt_row gt_center gt_striped'>Median (25%, 75%)</td>
#> <td class='gt_row gt_center gt_striped'>276 (196, 360)</td>
#> <td class='gt_row gt_center gt_striped'>120 (79, 160)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>hp</td>
#> <td class='gt_row gt_center'>Median (25%, 75%)</td>
#> <td class='gt_row gt_center'>175 (116, 192)</td>
#> <td class='gt_row gt_center'>109 (66, 113)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>drat</td>
#> <td class='gt_row gt_center gt_striped'>Median (25%, 75%)</td>
#> <td class='gt_row gt_center gt_striped'>3.15 (3.07, 3.70)</td>
#> <td class='gt_row gt_center gt_striped'>4.08 (3.85, 4.22)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>wt</td>
#> <td class='gt_row gt_center'>Median (25%, 75%)</td>
#> <td class='gt_row gt_center'>3.52 (3.44, 3.84)</td>
#> <td class='gt_row gt_center'>2.32 (1.94, 2.78)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>qsec</td>
#> <td class='gt_row gt_center gt_striped'>Median (25%, 75%)</td>
#> <td class='gt_row gt_center gt_striped'>17.82 (17.18, 19.17)</td>
#> <td class='gt_row gt_center gt_striped'>17.02 (16.46, 18.61)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>vs</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>7 (37%)</td>
#> <td class='gt_row gt_center'>7 (54%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>gear</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">3</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>15 (79%)</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">4</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>4 (21%)</td>
#> <td class='gt_row gt_center gt_striped'>8 (62%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">5</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>5 (38%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>carb</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">1</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>3 (16%)</td>
#> <td class='gt_row gt_center'>4 (31%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">2</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>6 (32%)</td>
#> <td class='gt_row gt_center gt_striped'>4 (31%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">3</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'>3 (16%)</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">4</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'>7 (37%)</td>
#> <td class='gt_row gt_center gt_striped'>3 (23%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">6</td>
#> <td class='gt_row gt_center'>n (%)</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>1 (7.7%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">8</td>
#> <td class='gt_row gt_center gt_striped'>n (%)</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>1 (7.7%)</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> </div>