as_flextable {gtsummary}	R Documentation

Convert gtsummary object to a flextable object

Description

Function converts a gtsummary object to a flextable object. A user can use this function if they wish to add customized formatting available via the flextable functions. The flextable output is particularly useful when combined with R markdown with Word output, since the gt package does not support Word.

Usage

as_flextable(x, ...)

## S3 method for class 'gtsummary'
as_flextable(
  x,
  include = everything(),
  return_calls = FALSE,
  strip_md_bold = TRUE,
  ...
)

Arguments

x
Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

...
Not used

include
Commands to include in output. Input may be a vector of quoted or unquoted names. tidyselect and gtsummary select helper functions are also accepted. Default is everything().

return_calls
Logical. Default is FALSE. If TRUE, the calls are returned as a list of expressions.

strip_md_bold
When TRUE, all double asterisk (markdown language for bold weight) in column labels and spanning headers are removed. Default is TRUE

Value

A flextable object

Details

The as_flextable() takes the data frame that will be printed and converts it to a flextable and formats the table with the following flextable functions.

flextable::flextable()

flextable::set_header_labels() to set column labels

flextable::add_header_row(), if applicable, to set spanning column header

flextable::align() to set column alignment

flextable::padding() to indent variable levels

flextable::fontsize() to set font size

flextable::autofit() to estimate the column widths

flextable::footnote() to add table footnotes and source notes

flextable::bold() to bold cells in data frame

flextable::italic() to italicize cells in data frame

Any one of these commands may be omitted using the include= argument.

Pro tip: Use the flextable::width() function for exacting control over column width after calling as_flextable().

Author(s)

Daniel D. Sjoberg

See Also

Other gtsummary output types: as_gt(), as_kable_extra(), as_kable(), as_tibble.gtsummary()

Examples

trial %>%
  select(trt, age, grade) %>%
  tbl_summary(by = trt) %>%
  add_p() %>%
  as_flextable()

[Package gtsummary version 1.3.2 Index]