Convert to gt_tbl object

Function converts object to a gt_tbl object. This function is used in the background when the results are printed or knit. A user can use this function if they wish to add customized formatting available via the gt package

as_gt(x, omit = NULL)

Arguments

x
object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

omit
vector of gt commands to omit. Default is NULL

Examples

trial %>% tbl_summary(by = "trt") %>% as_gt()

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
#> }
#> 
#> #zdnvsgovam .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #zdnvsgovam .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #zdnvsgovam .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #zdnvsgovam .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #zdnvsgovam .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #zdnvsgovam .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #zdnvsgovam .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #zdnvsgovam .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #zdnvsgovam .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #zdnvsgovam .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #zdnvsgovam .gt_striped tr:nth-child(even) {
#>   background-color: #f2f2f2;
#> }
#> 
#> #zdnvsgovam .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #zdnvsgovam .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #zdnvsgovam .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #zdnvsgovam .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #zdnvsgovam .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #zdnvsgovam .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #zdnvsgovam .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #zdnvsgovam .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #zdnvsgovam .gt_center {
#>   text-align: center;
#> }
#> 
#> #zdnvsgovam .gt_left {
#>   text-align: left;
#> }
#> 
#> #zdnvsgovam .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #zdnvsgovam .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #zdnvsgovam .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #zdnvsgovam .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #zdnvsgovam .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #zdnvsgovam .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="zdnvsgovam" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'>stat_1</th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'>stat_2</th>
#> </tr>
#> <tbody class='gt_table_body gt_striped'>
#> <tr>
#> <td class='gt_row gt_left'>Age, yrs</td>
#> <td class='gt_row gt_center'>47 (39, 58)</td>
#> <td class='gt_row gt_center'>46 (36, 54)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Unknown</td>
#> <td class='gt_row gt_center'>3</td>
#> <td class='gt_row gt_center'>5</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Marker Level, ng/mL</td>
#> <td class='gt_row gt_center'>0.61 (0.22, 1.20)</td>
#> <td class='gt_row gt_center'>0.72 (0.22, 1.63)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Unknown</td>
#> <td class='gt_row gt_center'>4</td>
#> <td class='gt_row gt_center'>4</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>T Stage</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T1</td>
#> <td class='gt_row gt_center'>25 (23%)</td>
#> <td class='gt_row gt_center'>26 (28%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T2</td>
#> <td class='gt_row gt_center'>26 (24%)</td>
#> <td class='gt_row gt_center'>23 (25%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T3</td>
#> <td class='gt_row gt_center'>29 (27%)</td>
#> <td class='gt_row gt_center'>13 (14%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T4</td>
#> <td class='gt_row gt_center'>27 (25%)</td>
#> <td class='gt_row gt_center'>31 (33%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Grade</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">I</td>
#> <td class='gt_row gt_center'>38 (36%)</td>
#> <td class='gt_row gt_center'>29 (31%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">II</td>
#> <td class='gt_row gt_center'>34 (32%)</td>
#> <td class='gt_row gt_center'>24 (26%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">III</td>
#> <td class='gt_row gt_center'>35 (33%)</td>
#> <td class='gt_row gt_center'>40 (43%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Tumor Response</td>
#> <td class='gt_row gt_center'>52 (51%)</td>
#> <td class='gt_row gt_center'>30 (33%)</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Unknown</td>
#> <td class='gt_row gt_center'>6</td>
#> <td class='gt_row gt_center'>3</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> </div>