Convert to knitr_kable object

Function converts gtsummary objects to a knitr_kable objects. This function is used in the background when the results are printed or knit. A user can use this function if they wish to add customized formatting available via knitr::kable.

as_kable(x, include = NULL, exclude = NULL, ...)

Arguments

x
Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

include
Character vector naming kable commands to include in printing. Default is NULL, which utilizes all commands in x$kable_calls.

exclude
Character vector naming kable commands to exclude in printing. Default is NULL.

...
Additional arguments passed to knitr::kable

Value

A knitr_kable object

See also

Examples

trial %>%
  tbl_summary(by = trt) %>%
  as_kable()

#> Warning: Results will be printed using 'knitr::kable()' and do not 
#> support footers or spanning headers. 
#> For tables styled by the gt package, use the installation code below.
#> 'remotes::install_github("rstudio/gt")'
#> 
#> If you prefer to always use 'knitr::kable()', add the option
#> 'options(gtsummary.print_engine = "kable")' to your script
#> or in a user- or project-level startup file, '.Rprofile'.
#> This warning is displayed once per session.

#> 
#> 
#> |**Characteristic**     |**Drug**, N = 107 |**Placebo**, N = 93 |
#> |:----------------------|:-----------------|:-------------------|
#> |Age, yrs               |47 (39, 58)       |45 (36, 54)         |
#> |Unknown                |6                 |3                   |
#> |Marker Level, ng/mL    |0.60 (0.22, 1.21) |0.72 (0.22, 1.53)   |
#> |Unknown                |5                 |4                   |
#> |T Stage                |                  |                    |
#> |T1                     |25 (23%)          |26 (28%)            |
#> |T2                     |26 (24%)          |23 (25%)            |
#> |T3                     |29 (27%)          |13 (14%)            |
#> |T4                     |27 (25%)          |31 (33%)            |
#> |Grade                  |                  |                    |
#> |I                      |38 (36%)          |29 (31%)            |
#> |II                     |34 (32%)          |24 (26%)            |
#> |III                    |35 (33%)          |40 (43%)            |
#> |Tumor Response         |53 (51%)          |30 (34%)            |
#> |Unknown                |4                 |5                   |
#> |Patient Died           |49 (46%)          |58 (62%)            |
#> |Months to Death/Censor |24.0 (16.0, 24.0) |20.9 (13.8, 24.0)   |