as_tibbleS3 {gtsummary}	R Documentation

Convert gtsummary object to tibble

Description

Function converts gtsummary objects tibbles. The formatting stored in x$kable_calls is applied.

Usage

## S3 method for class 'tbl_summary'
as_tibble(x, include = NULL, exclude = NULL, col_labels = TRUE, ...)

## S3 method for class 'tbl_regression'
as_tibble(x, include = NULL, exclude = NULL, col_labels = TRUE, ...)

## S3 method for class 'tbl_uvregression'
as_tibble(x, include = NULL, exclude = NULL, col_labels = TRUE, ...)

## S3 method for class 'tbl_merge'
as_tibble(x, include = NULL, exclude = NULL, col_labels = TRUE, ...)

## S3 method for class 'tbl_stack'
as_tibble(x, include = NULL, exclude = NULL, col_labels = TRUE, ...)

## S3 method for class 'tbl_survival'
as_tibble(x, include = NULL, exclude = NULL, col_labels = TRUE, ...)

Arguments

x
Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

include
Character vector naming kable commands to include in printing. Default is NULL, which utilizes all commands in x$kable_calls.

exclude
Character vector naming kable commands to exclude in printing. Default is NULL.

col_labels
Logical argument adding column labels to output tibble. Default is TRUE.

...
Not used

Value

a tibble

Author(s)

Daniel D. Sjoberg

See Also

tbl_summary tbl_regression tbl_uvregression tbl_survival

Examples

tbl <-
  trial %>%
  tbl_summary(by = trt)

as_tibble(tbl)

# without column labels
as_tibble(tbl, col_names = FALSE)

[Package gtsummary version 1.2.4 Index]