assign_dichotomous_value {gtsummary}	R Documentation

For dichotomous data, returns that value that will be printed in table.

Description

For dichotomous data, returns that value that will be printed in table.

Usage

assign_dichotomous_value(data, variable, summary_type, class, value)

Arguments

data
data frame

variable
character variable name in data that will be tabulated

summary_type
the type of summary statistics that will be calculated

class
class of variable

Value

value that will be printed in table for dichotomous data

Author(s)

Daniel D. Sjoberg

[Package gtsummary version 1.0.0 Index]