assign_sort {gtsummary}	R Documentation

Assigns categorical variables sort type ("alphanumeric" or "frequency")

Description

Assigns categorical variables sort type ("alphanumeric" or "frequency")

Usage

assign_sort(variable, summary_type, sort)

Arguments

variable
variable name

summary_type
the type of variable ("continuous", "categorical", "dichotomous")

sort
named list indicating the type of sorting to perform. Default is NULL. The names of the list elements are variable names or '..categorical..' for assigning all variables of that type. If both a variable name and '..categorical..' are specified, the variable name takes precedent

Author(s)

Daniel D. Sjoberg

[Package gtsummary version 1.1.1 Index]