assign_test {gtsummary}	R Documentation

determine the appropriate test type given two variables

Description

determine the appropriate test type given two variables

Usage

assign_test(data, var, var_summary_type, by_var, test, group)

Arguments

data
input data set

var
categorical or continuous variable for which a test with by_var is desired

var_summary_type
summary_type from meta data

by_var
categorical variable

test
list of user defined statistical tests and corresponding variables

Value

most appropriate test as text of the test function

Author(s)

Daniel D. Sjoberg

[Package gtsummary version 1.0.0 Index]