calculate_pvalue {gtsummary}	R Documentation

This function takes in the meta-data and calculates an appropriate p-value

Description

This function takes in the meta-data and calculates an appropriate p-value

Usage

calculate_pvalue(data, variable, by, test, type, group)

Arguments

data
input data set

variable
categorical or continuous variable for which a test with by_var is desired

by
categorical variable

test
list of statistical tests from meta data

type
the type of variable, one of categorical or continuous, from the metadata

group
the group variable for clustered data

Value

a table of p-values for each variable

Author(s)

Emily C. Zabor

[Package gtsummary version 1.1.0 Index]