calculate_summary_stat {gtsummary}	R Documentation

This function takes in the meta data table, and calls the appropriate summarize function.

Description

This function takes in the meta data table, and calls the appropriate summarize function.

Usage

calculate_summary_stat(data, variable, by, summary_type, dichotomous_value,
  var_label, stat_display, digits, class, missing, missing_text, sort,
  row_percent)

Arguments

data
Data frame

variable
Character variable name in data that will be tabulated

by
Character variable name indata that Summary statistics for variable are stratified

summary_type
A list that includes specified summary types.

dichotomous_value
If the output is dichotomous, then this is the value

var_label
String label

stat_display
String that specifies the format of the displayed statistics. The syntax follows glue::glue() inputs with n, N, and p as input options. of the variable that will be displayed.

digits
integer indicating the number of decimal places to be used.

class
variable class. If class is NA, then all values are NA, and no summary statistics will be calculated.

missing
whether to include NA values in the table. missing controls if the table includes counts of NA values: the allowed values correspond to never ("no"), only if the count is positive ("ifany") and even for zero counts ("always"). Default is "ifany".

missing_text
String to display for count of missing observations.

sort
string indicating whether to sort categorical variables by 'alphanumeric' or 'frequency'

row_percent
Logical value indicating whether to calculate percentages within column to across rows

Author(s)

Daniel D. Sjoberg

[Package gtsummary version 1.0.0 Index]