continuous_digits_guess {gtsummary}	R Documentation

Guesses how many digits to use in rounding continuous variables or summary statistics

Description

Guesses how many digits to use in rounding continuous variables or summary statistics

Usage

continuous_digits_guess(data, variable, summary_type, class,
  digits = NULL)

Arguments

x
vector containing the values of a continuous variable. This can be raw data values or a vector of summary statistics themselves

Value

the rounded values

Author(s)

Emily Zabor, Daniel D. Sjoberg

[Package gtsummary version 1.0.0 Index]