df_by {gtsummary}	R Documentation

Simple utility function to get extract and calculate additional information about the 'by' variable in tbl_summary

Description

Given a dataset and the name of the 'by' variable, this function returns a data frame with unique levels of the by variable, the by variable ID, a character version of the levels, and the column name for each level in the tbl_summary output data frame.

Usage

df_by(data, by)

Arguments

data
data frame

by
character name of the by variable found in data

Author(s)

Daniel D. Sjoberg

[Package gtsummary version 1.1.0 Index]