gtsummary-package {gtsummary}	R Documentation

gtsummary: Presentation-Ready Data Summary and Analytic Result Tables

Description

Creates presentation-ready tables summarizing data sets, regression models, and more. The code to create the tables is concise and highly customizable. Data frames can be summarized with any function, e.g. mean(), median(), even user-written functions. Regression models are summarized and include the reference rows for categorical variables. Common regression models, such as logistic regression and Cox proportional hazards regression, are automatically identified and the tables are pre-filled with appropriate column headers.

Author(s)

Maintainer: Daniel D. Sjoberg danield.sjoberg@gmail.com

Authors:

Emily C. Zabor

Margie Hannum

Karissa Whiting

See Also

Useful links:

http://www.danieldsjoberg.com/gtsummary/

Report bugs at https://github.com/ddsjoberg/gtsummary/issues

[Package gtsummary version 1.1.0 Index]