inline_text.gtsummary {gtsummary}	R Documentation

Report statistics from summary tables inline

Description

Report statistics from summary tables inline

Usage

## S3 method for class 'gtsummary'
inline_text(x, variable, level = NULL, column = NULL, pattern = NULL, ...)

Arguments

x
gtsummary object

variable
Variable name of statistic to present

level
Level of the variable to display for categorical variables. Default is NULL

column
Column name to return from x$table_body.

pattern
String indicating the statistics to return. Uses glue::glue formatting. Default is NULL

...
Not used

column + pattern

Some gtsummary tables report multiple statistics in a single cell, e.g. "{mean} ({sd})" in tbl_summary() or tbl_svysummary(). We often need to report just the mean or the SD, and that can be accomplished by using both the column= and pattern= arguments. When both of these arguments are specified, the column argument selects the column to report statistics from, and the pattern argument specifies which statistics to report, e.g. inline_text(x, column = "stat_1", pattern = "{mean}") reports just the mean from a tbl_summary().

[Package gtsummary version 1.4.2.9012 Index]