inline_text.tbl_summary {gtsummary}	R Documentation

Report statistics from summary tables inline

Description

Takes an object with class tbl_summary, and the location of the statistic to report and returns the statistic for reporting inline in an R markdown document. Detailed examples in the tbl_summary vignette

Usage

## S3 method for class 'tbl_summary'
inline_text(x, variable, level = NULL,
  column = ifelse(is.null(x$by), "stat_0", stop("Must specify column")),
  pvalue_fun = function(x) style_pvalue(x, prepend_p = TRUE), ...)

Arguments

x
object created from tbl_summary

variable
variable name of statistic to present

level
level of the variable to display for categorical variables. Can also specify the 'Unknown' row. Default is NULL

column
column name to return from x$table_body. Can also pass the level of a by variable.

pvalue_fun
function to round and format p-values. Default is style_pvalue. The function must have a numeric vector input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, purrr::partial(style_pvalue, digits = 2)).

...
not used

Author(s)

Daniel D. Sjoberg

See Also

Other tbl_summary tools: add_comparison, add_n, add_overall, add_q.tbl_summary, add_stat_label, bold_italicize_labels_levels, cols_label_summary, tbl_summary

Examples

t1 <- tbl_summary(trial)
t2 <- tbl_summary(trial, by = "trt") %>% add_comparison()

inline_text(t1, variable = "age")
inline_text(t2, variable = "grade", level = "I", column = "Drug")
inline_text(t2, variable = "grade", column = "p.value")

[Package gtsummary version 1.0.0 Index]