modify_table_header {gtsummary}	R Documentation

Modify table_header

Description

Use modify_table_styling() instead.

Usage

modify_table_header(
  x,
  column,
  label = NULL,
  hide = NULL,
  align = NULL,
  missing_emdash = NULL,
  indent = NULL,
  text_interpret = NULL,
  bold = NULL,
  italic = NULL,
  fmt_fun = NULL,
  footnote_abbrev = NULL,
  footnote = NULL,
  spanning_header = NULL
)

Arguments

x
gtsummary object

column
vector or selector of columns in x$table_body

label
string of column label

hide
logical indicating whether to hide column from output

align
string indicating alignment of column, must be one of c("left", "right", "center")

missing_emdash
string that evaluates to logical identifying rows to include em-dash for missing values, e.g. "reference_row == TRUE"

indent
string that evaluates to logical identifying rows to indent

text_interpret
string, must be one of "gt::md" or "gt::html"

bold
string that evaluates to logical identifying rows to bold

italic
string that evaluates to logical identifying rows to italicize

fmt_fun
function that formats the statistics in the column

footnote_abbrev
string with abbreviation definition, e.g. "CI = Confidence Interval"

footnote
string with text for column footnote

spanning_header
string with text for spanning header

[Package gtsummary version 1.4.0.9006 Index]