plot {gtsummary}	R Documentation

Plot Regression Coefficients

Description

The plot() function extracts x$table_body and passes the it to GGally::ggcoef_plot() along with a formatting options.

Usage

## S3 method for class 'tbl_regression'
plot(x, remove_header_rows = TRUE, remove_reference_rows = FALSE, ...)

## S3 method for class 'tbl_uvregression'
plot(x, remove_header_rows = TRUE, remove_reference_rows = FALSE, ...)

Arguments

x
'tbl_regression' or 'tbl_uvregression' object

remove_header_rows
logical indicating whether to remove header rows for categorical variables. Default is TRUE

remove_reference_rows
logical indicating whether to remove reference rows for categorical variables. Default is FALSE.

...
arguments passed to GGally::ggcoef_plot(...)

Details

Value

a ggplot

Examples

if (requireNamespace("GGally")) {
  glm(response ~ marker + grade, trial, family = binomial) %>%
    tbl_regression(
      add_estimate_to_reference_rows = TRUE,
      exponentiate = TRUE
    ) %>%
    plot()
}

[Package gtsummary version 1.4.0.9006 Index]