Renames columns to values in a row of data frame

Renames columns to values in a row of data frame

row_to_name(data, row_n = 1, delete_row = TRUE)

Arguments

data
input data set

row_n
row number to take names from. Default is row 1

delete_row
logical argument whether to delete row after renaming. Default is TRUE

Examples

fmt_table1(trial, by = "trt") %>%
  purrr::pluck("table1") %>%
  row_to_name()

#> # A tibble: 16 x 5
#>    NA       header2 Variable            Drug              Placebo          
#>    <chr>    <chr>   <chr>               <chr>             <chr>            
#>  1 NA       header1 ""                  N = 107           N = 93           
#>  2 age      label   Age, yrs            47 (39, 58)       46 (36, 54)      
#>  3 age      missing Unknown             3                 5                
#>  4 marker   label   Marker Level, ng/mL 0.61 (0.22, 1.20) 0.72 (0.22, 1.63)
#>  5 marker   missing Unknown             4                 4                
#>  6 stage    label   T Stage             NA                NA               
#>  7 stage    level   T1                  25 (23%)          26 (28%)         
#>  8 stage    level   T2                  26 (24%)          23 (25%)         
#>  9 stage    level   T3                  29 (27%)          13 (14%)         
#> 10 stage    level   T4                  27 (25%)          31 (33%)         
#> 11 grade    label   Grade               NA                NA               
#> 12 grade    level   I                   38 (36%)          29 (31%)         
#> 13 grade    level   II                  34 (32%)          24 (26%)         
#> 14 grade    level   III                 35 (33%)          40 (43%)         
#> 15 response label   Tumor Response      52 (51%)          30 (33%)         
#> 16 response missing Unknown             6                 3