summarize_continuous {gtsummary}	R Documentation

Calculates and formats summary statistics for continuous data

Description

Calculates and formats summary statistics for continuous data

Usage

summarize_continuous(data, variable, by, digits, var_label, stat_display,
  missing, missing_text)

Arguments

data
data frame

variable
Character variable name in data that will be tabulated

by
Character variable name in data that Summary statistics for variable are stratified

digits
vector of integers (or single integer) indicating the number of decimal places to be used for rounding.

var_label
string label

stat_display
String that specifies the format of the displayed statistics. The syntax follows glue inputs with n, N, and p as input options.

missing
whether to include NA values in the table. missing controls if the table includes counts of NA values: the allowed values correspond to never ("no"), only if the count is positive ("ifany") and even for zero counts ("always"). Default is "ifany".

missing_text
String to display for count of missing observations.

Value

formatted summary statistics in a tibble.

Author(s)

Daniel D. Sjoberg

[Package gtsummary version 1.1.1 Index]