Syntax and Notation

Syntax and Notation

Selectors

The gtsummary package also utilizes selectors: selectors from the tidyselect package and custom selectors. Review their help files for details.

tidy selectors

everything(), all_of(), any_of(), starts_with(), ends_with(), contains(), matches(), num_range(), last_col()

gtsummary selectors

all_continuous(), all_categorical(), all_dichotomous(), all_continuous2(), all_tests(), all_stat_cols(), all_interaction(), all_intercepts(), all_contrasts()

Formula and List Selectors

Many arguments throughout the gtsummary package accept list and formula notation, e.g. tbl_summary(statistic=). Below enumerates a few tips and shortcuts for using the list and formulas.

List of Formulas

Typical usage includes a list of formulas, where the LHS is a variable name or a selector.

tbl_summary(statistic = list(age ~ "{mean}", all_categorical() ~ "{n}"))

Named List

You may also pass a named list; however, the tidyselect and gtsummary selectors are not supported with this syntax.

tbl_summary(statistic = list(age = "{mean}", response = "{n}"))

Hybrid Named List/List of Formulas

Pass a combination of formulas and named elements

tbl_summary(statistic = list(age = "{mean}", all_categorical() ~ "{n}"))

Shortcuts

You can pass a single formula, which is equivalent to passing the formula in a list.

tbl_summary(statistic = all_categorical() ~ "{n}")

As a shortcut to select all variables, you can omit the LHS of the formula. The two calls below are equivalent.

tbl_summary(statistic = ~"{n}")
tbl_summary(statistic = everything() ~ "{n}")

Combination Selectors

Selectors can be combined using the c() function.

tbl_summary(statistic = c(everything(), -grade) ~ "{n}")