Turn a regression model object into a markdown-ready tibble.

This function uses broom::tidy from the broom or broom.mixed packages to perform the initial model formatting. Review the tbl_regression vignette for detailed examples.

tbl_regression(x, exponentiate = FALSE, label = NULL,
  include = names(stats::model.frame(x)), show_yesno = NULL,
  conf.level = 0.95, intercept = FALSE, coef_fun = style_coef,
  pvalue_fun = style_pvalue)

Arguments

x
regression model object

exponentiate
logical argument passed directly to tidy function. Default is FALSE

label
list of labels to write in the output. list(age60 = "Age > 60")

include
names of variables to include in output. Default is all variables.

show_yesno
Vector of names of categorical and factor variables that are c("No", "Yes"), c("no", "yes"), or c("NO", "YES") default to dichotomous printing (i.e. only Yes shown). To force both levels to be shown include the column name in show_yesno, e.g. show_yesno = c("highgrade", "female")

conf.level
confidence level passed directly to tidy function. Default is 0.95.

intercept
logical argument indicates whether to include the intercept in the output. Default is FALSE

coef_fun
function to round and format beta coefficients. Default is style_coef

pvalue_fun
function to round and format p-values. Default is style_pvalue

Examples

mod1 <-
  lm(hp ~ mpg + factor(cyl), mtcars) %>%
  tbl_regression()

mod2 <-
  glm(response ~ age + grade + stage, trial, family = binomial(link = "logit")) %>%
  tbl_regression(exponentiate = TRUE)

library(lme4)

#> Loading required package: Matrix

mod_glmer <-
  glmer(am ~ hp + (1 | gear), mtcars, family = binomial) %>%
  tbl_regression(exponentiate = TRUE)