tbl_summary_ {gtsummary}	R Documentation

Standard evaluation version of tbl_summary()

Description

The 'by =' argument can be passed as a string, rather than with non-standard evaluation as in tbl_summary. Review the help file for tbl_summary fully documented options and arguments.

Usage

tbl_summary_(data, by = NULL, label = NULL, statistic = NULL,
  digits = NULL, type = NULL, value = NULL, missing = c("ifany",
  "always", "no"), missing_text = "Unknown", sort = NULL,
  percent = c("column", "row", "cell"), group = NULL)

Arguments

data
A data frame

by
A column name in data. Summary statistics will be calculated separately for each level of the by variable (e.g. by = trt). If NULL, summary statistics are calculated using all observations.

label
List of formulas specifying variables labels, e.g. list(vars(age) ~ "Age, yrs", vars(ptstage) ~ "Path T Stage"). If a variable's label is not specified here, the function will take the label attribute (attr(data$age, "label")). If attribute label is NULL, the variable name will be used.

statistic
List of formulas specifying types of summary statistics to display for each variable. The default is list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}%)"). See below for details.

digits
List of formulas specifying the number of decimal places to round continuous summary statistics. If not specified, tbl_summary guesses an appropriate number of decimals to round statistics. When multiple statistics are displayed for a single variable, supply a vector rather than an integer. For example, if the statistic being calculated is "{mean} ({sd})" and you want the mean rounded to 1 decimal place, and the SD to 2 use digits = list("age" ~ c(1, 2)).

type
List of formulas specifying variable types. Accepted values are c("continuous", "categorical", "dichotomous"), e.g. type = list(starts_with(age) ~ "continuous", "female" ~ "dichotomous"). If type not specified for a variable, the function will default to an appropriate summary type. See below for details.

value
List of formulas specifying the value to display for dichotomous variables. See below for details.

missing
Indicates whether to include counts of NA values in the table. Allowed values are "no" (never display NA values), "ifany" (only display if any NA values), and "always" (includes NA count row for all variables). Default is "ifany".

missing_text
String to display for count of missing observations. Default is "Unknown".

sort
List of formulas specifying the type of sorting to perform for categorical data. Options are frequency where results are sorted in descending order of frequency and alphanumeric, e.g. sort = list(everything() ~ "frequency")

percent
Indicates the type of percentage to return. Must be one of "column", "row", or "cell". Default is "column".

group
DEPRECATED. Migrated to add_p

[Package gtsummary version 1.2.3 Index]