Creates table of univariate regression results

The tbl_uvregression function arguments are similar to the tbl_regression arguments. Review the tbl_uvregression vignette for detailed examples.

tbl_uvregression(data, method, y, method.args = NULL,
  formula = "{y} ~ {.x}", exponentiate = FALSE, label = NULL,
  show_yesno = NULL, conf.level = 0.95, coef_fun = style_sigfig,
  pvalue_fun = style_pvalue)

Arguments

data
Data frame to be used in univariate regression modeling. Data frame includes the outcome variable(s) and the independent variables.

method
Regression method (e.g. lm, glm, coxph, and more).

y
model outcome as a string (e.g. y = 'recurrence' or y = 'Surv(time, recur)')

method.args
List of additional arguments passed on to the regression function defined by method.

formula
String that becomes the model formula. Uses glue syntax. Default is "{y} ~ {.x}", where {y} is the dependent variable, and {.x} represents a single covariate. For a random intercept, the formula may be formula = "{y} ~ {.x} + (1 | gear)".

exponentiate
logical argument passed directly to broom::tidy(). Default is FALSE

label
list of labels to write in the output. list(age60 = "Age > 60")

show_yesno
Vector of names of categorical and factor variables that are c("No", "Yes"), c("no", "yes"), or c("NO", "YES") default to dichotomous printing (i.e. only Yes shown). To force both levels to be shown include the column name in show_yesno, e.g. show_yesno = c("highgrade", "female")

conf.level
confidence level passed directly to broom::tidy. Default is 0.95.

coef_fun
function to round and format beta coefficients. Default is style_sigfig

pvalue_fun
function to round and format p-values. Default is style_pvalue

Examples

tbl_uvregression(
  trial,
  method = "glm",
  y = "response",
  method.args = list(family = binomial),
  exponentiate = TRUE
)

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
#> }
#> 
#> #erwyoprgbx .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #erwyoprgbx .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #erwyoprgbx .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #erwyoprgbx .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #erwyoprgbx .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #erwyoprgbx .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #erwyoprgbx .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #erwyoprgbx .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #erwyoprgbx .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #erwyoprgbx .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #erwyoprgbx .gt_striped tr:nth-child(even) {
#>   background-color: #f2f2f2;
#> }
#> 
#> #erwyoprgbx .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #erwyoprgbx .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #erwyoprgbx .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #erwyoprgbx .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #erwyoprgbx .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #erwyoprgbx .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #erwyoprgbx .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #erwyoprgbx .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #erwyoprgbx .gt_center {
#>   text-align: center;
#> }
#> 
#> #erwyoprgbx .gt_left {
#>   text-align: left;
#> }
#> 
#> #erwyoprgbx .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #erwyoprgbx .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #erwyoprgbx .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #erwyoprgbx .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #erwyoprgbx .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #erwyoprgbx .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="erwyoprgbx" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>N</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Coefficient</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>95% Confidence Interval</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>p-value</strong></th>
#> </tr>
#> <tbody class='gt_table_body gt_striped'>
#> <tr>
#> <td class='gt_row gt_left'>Treatment Randomization</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Drug</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Placebo</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.47</td>
#> <td class='gt_row gt_center'>0.26, 0.84</td>
#> <td class='gt_row gt_center'>0.012</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Age, yrs</td>
#> <td class='gt_row gt_center'>183</td>
#> <td class='gt_row gt_center'>1.00</td>
#> <td class='gt_row gt_center'>0.98, 1.0</td>
#> <td class='gt_row gt_center'>>0.9</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Marker Level, ng/mL</td>
#> <td class='gt_row gt_center'>183</td>
#> <td class='gt_row gt_center'>0.92</td>
#> <td class='gt_row gt_center'>0.65, 1.3</td>
#> <td class='gt_row gt_center'>0.6</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>T Stage</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T1</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T2</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>1.4</td>
#> <td class='gt_row gt_center'>0.63, 3.1</td>
#> <td class='gt_row gt_center'>0.4</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T3</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.63</td>
#> <td class='gt_row gt_center'>0.26, 1.5</td>
#> <td class='gt_row gt_center'>0.3</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T4</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.50</td>
#> <td class='gt_row gt_center'>0.22, 1.1</td>
#> <td class='gt_row gt_center'>0.089</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Grade</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">I</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">II</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.60</td>
#> <td class='gt_row gt_center'>0.29, 1.3</td>
#> <td class='gt_row gt_center'>0.2</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">III</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.76</td>
#> <td class='gt_row gt_center'>0.38, 1.5</td>
#> <td class='gt_row gt_center'>0.4</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> </div>

# rounding pvalues to 2 decimal places, and adding global p-values
tbl_uvregression(
  trial,
  method = "glm",
  y = "response",
  method.args = list(family = binomial),
  exponentiate = TRUE,
  pvalue_fun = function(x) style_pvalue(x, digits = 2)
) %>%
  add_global()

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
#> }
#> 
#> #bzpexapgxv .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #bzpexapgxv .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #bzpexapgxv .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #bzpexapgxv .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #bzpexapgxv .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #bzpexapgxv .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #bzpexapgxv .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #bzpexapgxv .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #bzpexapgxv .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #bzpexapgxv .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #bzpexapgxv .gt_striped tr:nth-child(even) {
#>   background-color: #f2f2f2;
#> }
#> 
#> #bzpexapgxv .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #bzpexapgxv .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #bzpexapgxv .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #bzpexapgxv .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #bzpexapgxv .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #bzpexapgxv .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #bzpexapgxv .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #bzpexapgxv .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #bzpexapgxv .gt_center {
#>   text-align: center;
#> }
#> 
#> #bzpexapgxv .gt_left {
#>   text-align: left;
#> }
#> 
#> #bzpexapgxv .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #bzpexapgxv .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #bzpexapgxv .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #bzpexapgxv .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #bzpexapgxv .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #bzpexapgxv .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="bzpexapgxv" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>N</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Coefficient</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>95% Confidence Interval</strong></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>p-value</strong></th>
#> </tr>
#> <tbody class='gt_table_body gt_striped'>
#> <tr>
#> <td class='gt_row gt_left'>Treatment Randomization</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.011</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Drug</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Placebo</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.47</td>
#> <td class='gt_row gt_center'>0.26, 0.84</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Age, yrs</td>
#> <td class='gt_row gt_center'>183</td>
#> <td class='gt_row gt_center'>1.00</td>
#> <td class='gt_row gt_center'>0.98, 1.0</td>
#> <td class='gt_row gt_center'>0.90</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Marker Level, ng/mL</td>
#> <td class='gt_row gt_center'>183</td>
#> <td class='gt_row gt_center'>0.92</td>
#> <td class='gt_row gt_center'>0.65, 1.3</td>
#> <td class='gt_row gt_center'>0.64</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>T Stage</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.055</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T1</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T2</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>1.4</td>
#> <td class='gt_row gt_center'>0.63, 3.1</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T3</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.63</td>
#> <td class='gt_row gt_center'>0.26, 1.5</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T4</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.50</td>
#> <td class='gt_row gt_center'>0.22, 1.1</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Grade</td>
#> <td class='gt_row gt_center'>191</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.40</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">I</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'>—</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">II</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.60</td>
#> <td class='gt_row gt_center'>0.29, 1.3</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">III</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.76</td>
#> <td class='gt_row gt_center'>0.38, 1.5</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> </div>