Available gtsummary themes

The following themes are available to use within the gtsummary package. Print theme elements with theme_gtsummary_journal(set_theme = FALSE |> print(). Review the themes vignette for details.

Usage

theme_gtsummary_journal(
  journal = c("jama", "lancet", "nejm", "qjecon"),
  set_theme = TRUE
)

theme_gtsummary_compact(set_theme = TRUE, font_size = NULL)

theme_gtsummary_printer(
  print_engine = c("gt", "kable", "kable_extra", "flextable", "huxtable", "tibble"),
  set_theme = TRUE
)

theme_gtsummary_language(
  language = c("de", "en", "es", "fr", "gu", "hi", "is", "ja", "kr", "mr", "nl", "no",
    "pt", "se", "zh-cn", "zh-tw"),
  decimal.mark = NULL,
  big.mark = NULL,
  iqr.sep = NULL,
  ci.sep = NULL,
  set_theme = TRUE
)

theme_gtsummary_continuous2(
  statistic = "{median} ({p25, {p75})",
  set_theme = TRUE
)

theme_gtsummary_mean_sd(set_theme = TRUE)

theme_gtsummary_eda(set_theme = TRUE)

Arguments

journal

String indicating the journal theme to follow. One of c("jama", "lancet", "nejm", "qjecon"). Details below.

set_theme

(scalar logical)
Logical indicating whether to set the theme. Default is TRUE. When FALSE the named list of theme elements is returned invisibly

font_size

(scaler numeric)
Numeric font size for compact theme. Default is 13 for gt tables, and 8 for all other output types

print_engine

String indicating the print method. Must be one of "gt", "kable", "kable_extra", "flextable", "tibble"

language

(string)
String indicating language. Must be one of "de" (German), "en" (English), "es" (Spanish), "fr" (French), "gu" (Gujarati), "hi" (Hindi), "is" (Icelandic),"ja" (Japanese), "kr" (Korean), "nl" (Dutch), "mr" (Marathi), "no" (Norwegian), "pt" (Portuguese), "se" (Swedish), "zh-cn" (Chinese Simplified), "zh-tw" (Chinese Traditional)

If a language is missing a translation for a word or phrase, please feel free to reach out on GitHub with the translated text.

decimal.mark

(string)
The character to be used to indicate the numeric decimal point. Default is "." or getOption("OutDec")

big.mark

(string)
Character used between every 3 digits to separate hundreds/thousands/millions/etc. Default is ",", except when decimal.mark = "," when the default is a space.

iqr.sep

(string)
String indicating separator for the default IQR in tbl_summary(). If decimal.mark= is NULL, iqr.sep= is ", ". The comma separator, however, can look odd when decimal.mark = ",". In this case the argument will default to an en dash

ci.sep

(string)
String indicating separator for confidence intervals. If decimal.mark= is NULL, ci.sep= is ", ". The comma separator, however, can look odd when decimal.mark = ",". In this case the argument will default to an en dash

statistic

Default statistic continuous variables

Themes

theme_gtsummary_journal(journal)

"jama" The Journal of the American Medical Association

Round large p-values to 2 decimal places; separate confidence intervals with "ll to ul".

tbl_summary() Doesn't show percent symbol; use em-dash to separate IQR; run add_stat_label()

tbl_regression()/tbl_uvregression() show coefficient and CI in same column

"lancet" The Lancet

Use mid-point as decimal separator; round large p-values to 2 decimal places; separate confidence intervals with "ll to ul".

tbl_summary() Doesn't show percent symbol; use em-dash to separate IQR

"nejm" The New England Journal of Medicine

Round large p-values to 2 decimal places; separate confidence intervals with "ll to ul".

tbl_summary() Doesn't show percent symbol; use em-dash to separate IQR

"qjecon" The Quarterly Journal of Economics

tbl_summary() all percentages rounded to one decimal place

tbl_regression()/tbl_uvregression() add significance stars with add_significance_stars(); hides CI and p-value from output

theme_gtsummary_compact()

tables printed with gt, flextable, kableExtra, or huxtable will be compact with smaller font size and reduced cell padding

theme_gtsummary_printer(print_engine)

Use this theme to permanently change the default printer.

theme_gtsummary_continuous2()

Set all continuous variables to summary type "continuous2" by default

theme_gtsummary_mean_sd()

Set default summary statistics to mean and standard deviation in tbl_summary()

Set default continuous tests in add_p() to t-test and ANOVA

theme_gtsummary_eda()

Set all continuous variables to summary type "continuous2" by default

In tbl_summary() show the median, mean, IQR, SD, and Range by default

Use reset_gtsummary_theme() to restore the default settings

Review the themes vignette to create your own themes.

See also

Examples

# Setting JAMA theme for gtsummary
theme_gtsummary_journal("jama")
#> Setting theme "JAMA"
# Themes can be combined by including more than one
theme_gtsummary_compact()
#> Setting theme "Compact"

trial |>
  select(age, grade, trt) |>
  tbl_summary(by = trt) |>
  as_gt()
#> <div id="vilaicgibn" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#vilaicgibn table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #vilaicgibn thead, #vilaicgibn tbody, #vilaicgibn tfoot, #vilaicgibn tr, #vilaicgibn td, #vilaicgibn th {
#>   border-style: none;
#> }
#> 
#> #vilaicgibn p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #vilaicgibn .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 13px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #vilaicgibn .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #vilaicgibn .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #vilaicgibn .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #vilaicgibn .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #vilaicgibn .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #vilaicgibn .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #vilaicgibn .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #vilaicgibn .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #vilaicgibn .gt_group_heading {
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #vilaicgibn .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #vilaicgibn .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #vilaicgibn .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #vilaicgibn .gt_row {
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #vilaicgibn .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vilaicgibn .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #vilaicgibn .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #vilaicgibn .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #vilaicgibn .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vilaicgibn .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #vilaicgibn .gt_last_summary_row {
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vilaicgibn .gt_first_grand_summary_row {
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_last_grand_summary_row_top {
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #vilaicgibn .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vilaicgibn .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vilaicgibn .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vilaicgibn .gt_left {
#>   text-align: left;
#> }
#> 
#> #vilaicgibn .gt_center {
#>   text-align: center;
#> }
#> 
#> #vilaicgibn .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #vilaicgibn .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #vilaicgibn .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #vilaicgibn .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #vilaicgibn .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #vilaicgibn .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #vilaicgibn .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #vilaicgibn .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #vilaicgibn .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #vilaicgibn .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #vilaicgibn .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #vilaicgibn .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="&lt;strong&gt;Characteristic&lt;/strong&gt;"><strong>Characteristic</strong></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="&lt;strong&gt;Drug A&lt;/strong&gt;&lt;br /&gt;&#10;N = 98"><strong>Drug A</strong><br />
#> N = 98</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="&lt;strong&gt;Drug B&lt;/strong&gt;&lt;br /&gt;&#10;N = 102"><strong>Drug B</strong><br />
#> N = 102</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="label" class="gt_row gt_left">Age, Median (IQR)</td>
#> <td headers="stat_1" class="gt_row gt_center">46 (37 – 60)</td>
#> <td headers="stat_2" class="gt_row gt_center">48 (39 – 56)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    Unknown</td>
#> <td headers="stat_1" class="gt_row gt_center">7</td>
#> <td headers="stat_2" class="gt_row gt_center">4</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">Grade, n (%)</td>
#> <td headers="stat_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    I</td>
#> <td headers="stat_1" class="gt_row gt_center">35 (36)</td>
#> <td headers="stat_2" class="gt_row gt_center">33 (32)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    II</td>
#> <td headers="stat_1" class="gt_row gt_center">32 (33)</td>
#> <td headers="stat_2" class="gt_row gt_center">36 (35)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    III</td>
#> <td headers="stat_1" class="gt_row gt_center">31 (32)</td>
#> <td headers="stat_2" class="gt_row gt_center">33 (32)</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# reset gtsummary themes
reset_gtsummary_theme()