tidy_vetted {gtsummary}	R Documentation

Vetted tidy models

Description

Below is a list of models vetted for use in tbl_regression and tbl_uvregression. If a model is passed to these functions and the model is not listed below and a tidy() function is not specified in the tidy_fun= argument, the model object will be passed to broom::tidy.

stats::lm

stats::glm

survival::coxph

survival::survreg

lme4::glmer

lme4::lmer

geepack::geeglm

model support

If broom::tidy or broom.mixed::tidy support a class of model not listed above, please submit a GitHub Issue. The model can be added to the list of vetted models and unit tests will be put in place to ensure continued support for the model.

[Package gtsummary version 1.2.4 Index]