Convert character vector to data frame

This is used in some of the selecting we allow for, for example in as_gt(include=) you can use tidyselect to select among the call names to include.

Usage

vec_to_df(x)

Arguments

x

character vector

Value

data frame