Adds p-values to the output comparing values across groups

Adds p-values to the output comparing values across groups

add_comparison(x, test = NULL, pvalue_fun = style_pvalue,
  group = x$inputs$group)

Arguments

x
object with class tbl_summary from the tbl_summary function

test
user defined list of statistical tests provided as a named character vector with variables as names and test functions as values., e.g. list(age = "t.test", ptstage = "fisher.test"). Options include "t.test" for a T-test, "wilcox.test" for a Wilcoxon rank sum test, "kruskal.test" for a Kruskal-Wallis rank sum test, "chisq.test" for a Chi-squared test, "fisher.test" for a Fisher's exact test, and "re" for a random intercept model to account for clustered data. For "re" to be used "group" must also be specified in the function call.

pvalue_fun
function for rounding/formatting p-values. Default is style_pvalue. The function must have a single input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, partial(style_pvalue, digits = 2)).

group
Character vector of an ID or grouping variable. Summary statistics will not be printed for this column, but they may be used in subsequent functions. For example, the group column may be used in add_comparison() to include p-values with correlated data. Default is the group = input from tbl_summary

Examples

trial %>% tbl_summary(by = "trt") %>% add_comparison()

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #cvpeammhtu .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: auto;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> #cvpeammhtu .gt_heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#>   border-bottom-color: #FFFFFF;
#> }
#> 
#> #cvpeammhtu .gt_title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#>   padding-bottom: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #cvpeammhtu .gt_subtitle {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.subtitle.font.size */
#>   padding-top: 1px;
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #cvpeammhtu .gt_bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> #cvpeammhtu .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #cvpeammhtu .gt_col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* column_labels.background.color */
#>   font-size: 16px;
#>   /* column_labels.font.size */
#>   font-weight: initial;
#>   /* column_labels.font.weight */
#>   vertical-align: middle;
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> #cvpeammhtu .gt_sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> #cvpeammhtu .gt_group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #cvpeammhtu .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#>   vertical-align: middle;
#> }
#> 
#> #cvpeammhtu .gt_striped {
#>   background-color: #f2f2f2;
#> }
#> 
#> #cvpeammhtu .gt_row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#>   vertical-align: middle;
#> }
#> 
#> #cvpeammhtu .gt_stub {
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#>   padding-left: 12px;
#> }
#> 
#> #cvpeammhtu .gt_stub.gt_row {
#>   background-color: #FFFFFF;
#> }
#> 
#> #cvpeammhtu .gt_summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> #cvpeammhtu .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> #cvpeammhtu .gt_table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> #cvpeammhtu .gt_footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> #cvpeammhtu .gt_sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> #cvpeammhtu .gt_center {
#>   text-align: center;
#> }
#> 
#> #cvpeammhtu .gt_left {
#>   text-align: left;
#> }
#> 
#> #cvpeammhtu .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #cvpeammhtu .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #cvpeammhtu .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #cvpeammhtu .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #cvpeammhtu .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #cvpeammhtu .gt_footnote_glyph {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="cvpeammhtu" style="overflow-x:auto;"><!--gt table start-->
#> <table class='gt_table'>
#> <tr>
#> <th class='gt_col_heading gt_left' rowspan='1' colspan='1'><strong>Characteristic</strong><sup class='gt_footnote_glyph'>1</sup></th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Drug</strong>, N = 107</th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>Placebo</strong>, N = 93</th>
#> <th class='gt_col_heading gt_center' rowspan='1' colspan='1'><strong>p-value</strong><sup class='gt_footnote_glyph'>2</sup></th>
#> </tr>
#> <tbody class='gt_table_body'>
#> <tr>
#> <td class='gt_row gt_left'>Age, yrs</td>
#> <td class='gt_row gt_center'>47 (39, 58)</td>
#> <td class='gt_row gt_center'>46 (36, 54)</td>
#> <td class='gt_row gt_center'>0.3</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">Unknown</td>
#> <td class='gt_row gt_center gt_striped'>3</td>
#> <td class='gt_row gt_center gt_striped'>5</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>Marker Level, ng/mL</td>
#> <td class='gt_row gt_center'>0.61 (0.22, 1.20)</td>
#> <td class='gt_row gt_center'>0.72 (0.22, 1.63)</td>
#> <td class='gt_row gt_center'>0.4</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">Unknown</td>
#> <td class='gt_row gt_center gt_striped'>4</td>
#> <td class='gt_row gt_center gt_striped'>4</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left'>T Stage</td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'></td>
#> <td class='gt_row gt_center'>0.13</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">T1</td>
#> <td class='gt_row gt_center gt_striped'>25 (23%)</td>
#> <td class='gt_row gt_center gt_striped'>26 (28%)</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T2</td>
#> <td class='gt_row gt_center'>26 (24%)</td>
#> <td class='gt_row gt_center'>23 (25%)</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">T3</td>
#> <td class='gt_row gt_center gt_striped'>29 (27%)</td>
#> <td class='gt_row gt_center gt_striped'>13 (14%)</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">T4</td>
#> <td class='gt_row gt_center'>27 (25%)</td>
#> <td class='gt_row gt_center'>31 (33%)</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>Grade</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'></td>
#> <td class='gt_row gt_center gt_striped'>0.3</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">I</td>
#> <td class='gt_row gt_center'>38 (36%)</td>
#> <td class='gt_row gt_center'>29 (31%)</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped' style="text-align:left;text-indent:10px;">II</td>
#> <td class='gt_row gt_center gt_striped'>34 (32%)</td>
#> <td class='gt_row gt_center gt_striped'>24 (26%)</td>
#> <td class='gt_row gt_center gt_striped'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">III</td>
#> <td class='gt_row gt_center'>35 (33%)</td>
#> <td class='gt_row gt_center'>40 (43%)</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left gt_striped'>Tumor Response</td>
#> <td class='gt_row gt_center gt_striped'>52 (51%)</td>
#> <td class='gt_row gt_center gt_striped'>30 (33%)</td>
#> <td class='gt_row gt_center gt_striped'>0.017</td>
#> </tr>
#> <tr>
#> <td class='gt_row gt_left' style="text-align:left;text-indent:10px;">Unknown</td>
#> <td class='gt_row gt_center'>6</td>
#> <td class='gt_row gt_center'>3</td>
#> <td class='gt_row gt_center'></td>
#> </tr>
#> </tbody>
#> <tfoot>
#> <tr>
#> <td colspan='4' class='gt_footnote'><sup class='gt_footnote_glyph'><em>1</em></sup> Statistics presented: median (IQR) for continuous variables, n (%) for categorical variables<br /><sup class='gt_footnote_glyph'><em>2</em></sup> Statistical tests performed: Wilcoxon signed-rank test, chi-square test</td>
#> </tr>
#> </tfoot></table>
#> <!--gt table end-->
#> </div>