Convert to gt_tbl object

Function converts object to a gt_tbl object. This function is used in the background when the results are printed or knit. A user can use this function if they wish to add customized formatting available via the gt package

as_gt(x, omit = NULL)

Arguments

x
object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

omit
vector of gt commands to omit. Default is NULL

See also

Examples

tbl_gt <- trial %>% tbl_summary(by = "trt") %>% as_gt()