Convert gtsummary object to a huxtable object

Function converts a gtsummary object to a huxtable object. A user can use this function if they wish to add customized formatting available via the huxtable functions. The huxtable package supports output to PDF via LaTeX, as well as HTML and Word.

as_hux_table(
  x,
  include = everything(),
  return_calls = FALSE,
  strip_md_bold = FALSE
)

as_hux_xlsx(x, file, include = everything(), bold_header_rows = TRUE)

Arguments

x

Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

include

Commands to include in output. Input may be a vector of quoted or unquoted names. tidyselect and gtsummary select helper functions are also accepted. Default is everything().

return_calls

Logical. Default is FALSE. If TRUE, the calls are returned as a list of expressions.

strip_md_bold

DEPRECATED

file

File path for the output.

bold_header_rows

logical indicating whether to bold header rows. Default is TRUE

Value

A huxtable object

Excel Output

Use the as_hux_xlsx() function to save a copy of the table in an excel file. The file is saved using huxtable::quick_xlsx().

See also

Author

David Hugh-Jones, Daniel D. Sjoberg

Examples

# \donttest{
trial %>%
  dplyr::select(trt, age, grade) %>%
  tbl_summary(by = trt) %>%
  add_p() %>%
  as_hux_table()
#>           Characteristic   Drug A, N = 98   Drug B, N = 102   p-value  
#>         ───────────────────────────────────────────────────────────────
#>           Age               46 (37, 59)       48 (39, 56)       0.7    
#>           Unknown                7                 4                   
#>           Grade                                                 0.9    
#>           I                   35 (36%)         33 (32%)                
#>           II                  32 (33%)         36 (35%)                
#>           III                 31 (32%)         33 (32%)                
#>         ───────────────────────────────────────────────────────────────
#>           Median (IQR); n (%)                                          
#>           Wilcoxon rank sum test; Pearson's Chi-squared test           
#> 
#> Column names: label, stat_1, stat_2, p.value
# }