as_kable {gtsummary}	R Documentation

Convert gtsummary object to a kable object

Description

Function converts a gtsummary object to a knitr_kable object. This function is used in the background when the results are printed or knit. A user can use this function if they wish to add customized formatting available via knitr::kable.

Output from knitr::kable is less full featured compared to summary tables produced with gt. For example, kable summary tables do not include indentation, footnotes, or spanning header rows.

Usage

as_kable(x, include = everything(), return_calls = FALSE, exclude = NULL, ...)

Arguments

x
Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

include
Commands to include in output. Input may be a vector of quoted or unquoted names. tidyselect and gtsummary select helper functions are also accepted. Default is everything().

return_calls
Logical. Default is FALSE. If TRUE, the calls are returned as a list of expressions.

exclude
DEPRECATED

...
Additional arguments passed to knitr::kable

Details

Tip: To better distinguish variable labels and level labels when indenting is not supported, try bold_labels() or italicize_levels().

Value

A knitr_kable object

Author(s)

Daniel D. Sjoberg

See Also

Other gtsummary output types: as_flex_table(), as_gt(), as_hux_table(), as_kable_extra(), as_tibble.gtsummary()

Examples

trial %>%
  tbl_summary(by = trt) %>%
  bold_labels() %>%
  as_kable()

[Package gtsummary version 1.4.2.9012 Index]