Convert gtsummary object to a kable object

Function converts a gtsummary object to a knitr_kable object. This function is used in the background when the results are printed or knit. A user can use this function if they wish to add customized formatting available via knitr::kable.

Output from knitr::kable is less full featured compared to summary tables produced with gt. For example, kable summary tables do not include indentation, footnotes, or spanning header rows.

as_kable(x, include = everything(), return_calls = FALSE, exclude = NULL, ...)

Arguments

x
Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

include
Commands to include in output. Input may be a vector of quoted or unquoted names. tidyselect and gtsummary select helper functions are also accepted. Default is everything(), which includes all commands in x$kable_calls.

return_calls
Logical. Default is FALSE. If TRUE, the calls are returned as a list of expressions.

exclude
DEPRECATED

...
Additional arguments passed to knitr::kable

Value

A knitr_kable object

Details

Tip: To better distinguish variable labels and level labels when indenting is not supported, try bold_labels() or italicize_levels().

See also

Examples

trial %>%
  tbl_summary(by = trt) %>%
  bold_labels() %>%
  as_kable()

#> 
#> 
#> |**Characteristic**         |**Drug A**, N = 98 |**Drug B**, N = 102 |
#> |:--------------------------|:------------------|:-------------------|
#> |__Age, yrs__               |46 (37, 59)        |48 (39, 56)         |
#> |__Unknown__                |7                  |4                   |
#> |__Marker Level, ng/mL__    |0.84 (0.24, 1.57)  |0.52 (0.19, 1.20)   |
#> |__Unknown__                |6                  |4                   |
#> |__T Stage__                |                   |                    |
#> |__T1__                     |28 (29%)           |25 (25%)            |
#> |__T2__                     |25 (26%)           |29 (28%)            |
#> |__T3__                     |22 (22%)           |21 (21%)            |
#> |__T4__                     |23 (23%)           |27 (26%)            |
#> |__Grade__                  |                   |                    |
#> |__I__                      |35 (36%)           |33 (32%)            |
#> |__II__                     |32 (33%)           |36 (35%)            |
#> |__III__                    |31 (32%)           |33 (32%)            |
#> |__Tumor Response__         |28 (29%)           |33 (34%)            |
#> |__Unknown__                |3                  |4                   |
#> |__Patient Died__           |52 (53%)           |60 (59%)            |
#> |__Months to Death/Censor__ |23.5 (17.4, 24.0)  |21.2 (14.6, 24.0)   |