Convert gtsummary object to tibble

Function converts gtsummary objects tibbles. The formatting stored in x$kable_calls is applied.

# S3 method for gtsummary
as_tibble(
  x,
  include = everything(),
  col_labels = TRUE,
  return_calls = FALSE,
  exclude = NULL,
  ...
)

Arguments

x
Object created by a function from the gtsummary package (e.g. tbl_summary or tbl_regression)

include
Commands to include in output. Input may be a vector of quoted or unquoted names. tidyselect and gtsummary select helper functions are also accepted. Default is everything(), which includes all commands in x$kable_calls.

col_labels
Logical argument adding column labels to output tibble. Default is TRUE.

return_calls
Logical. Default is FALSE. If TRUE, the calls are returned as a list of expressions.

exclude
DEPRECATED

...
Not used

Value

a tibble

See also

Examples

tbl <-
  trial %>%
  tbl_summary(by = trt)

as_tibble(tbl)

#> # A tibble: 17 x 3
#>    `**Characteristic**`   `**Drug A**, N = 98` `**Drug B**, N = 102`
#>    <chr>                  <chr>                <chr>                
#>  1 Age, yrs               46 (37, 59)          48 (39, 56)          
#>  2 Unknown                7                    4                    
#>  3 Marker Level, ng/mL    0.84 (0.24, 1.57)    0.52 (0.19, 1.20)    
#>  4 Unknown                6                    4                    
#>  5 T Stage                NA                   NA                   
#>  6 T1                     28 (29%)             25 (25%)             
#>  7 T2                     25 (26%)             29 (28%)             
#>  8 T3                     22 (22%)             21 (21%)             
#>  9 T4                     23 (23%)             27 (26%)             
#> 10 Grade                  NA                   NA                   
#> 11 I                      35 (36%)             33 (32%)             
#> 12 II                     32 (33%)             36 (35%)             
#> 13 III                    31 (32%)             33 (32%)             
#> 14 Tumor Response         28 (29%)             33 (34%)             
#> 15 Unknown                3                    4                    
#> 16 Patient Died           52 (53%)             60 (59%)             
#> 17 Months to Death/Censor 23.5 (17.4, 24.0)    21.2 (14.6, 24.0)    

# without column labels
as_tibble(tbl, col_names = FALSE)

#> # A tibble: 17 x 3
#>    `**Characteristic**`   `**Drug A**, N = 98` `**Drug B**, N = 102`
#>    <chr>                  <chr>                <chr>                
#>  1 Age, yrs               46 (37, 59)          48 (39, 56)          
#>  2 Unknown                7                    4                    
#>  3 Marker Level, ng/mL    0.84 (0.24, 1.57)    0.52 (0.19, 1.20)    
#>  4 Unknown                6                    4                    
#>  5 T Stage                NA                   NA                   
#>  6 T1                     28 (29%)             25 (25%)             
#>  7 T2                     25 (26%)             29 (28%)             
#>  8 T3                     22 (22%)             21 (21%)             
#>  9 T4                     23 (23%)             27 (26%)             
#> 10 Grade                  NA                   NA                   
#> 11 I                      35 (36%)             33 (32%)             
#> 12 II                     32 (33%)             36 (35%)             
#> 13 III                    31 (32%)             33 (32%)             
#> 14 Tumor Response         28 (29%)             33 (34%)             
#> 15 Unknown                3                    4                    
#> 16 Patient Died           52 (53%)             60 (59%)             
#> 17 Months to Death/Censor 23.5 (17.4, 24.0)    21.2 (14.6, 24.0)