gtsummary table dimension

Returns the dimension of a gtsummary table, that is, the number of rows and the number of un-hidden columns.

nrow() calls dim(); therefore, nrow() will also work on gtsummary tables.

Usage

# S3 method for class 'gtsummary'
dim(x)

Arguments

x

(gtsummary)
a 'gtsummary' table

Value

integer vector

Examples

tbl <- tbl_summary(trial, include = age, by = trt)

dim(tbl)
#> [1] 2 3
nrow(tbl)
#> [1] 2