gtsummary: Presentation-Ready Data Summary and Analytic Result Tables

Creates presentation-ready tables summarizing data sets, regression models, and more. The code to create the tables is concise and highly customizable. Data frames can be summarized with any function, e.g. mean(), median(), even user-written functions. Regression models are summarized and include the reference rows for categorical variables. Common regression models, such as logistic regression and Cox proportional hazards regression, are automatically identified and the tables are pre-filled with appropriate column headers.

See also

Author

Maintainer: Daniel D. Sjoberg danield.sjoberg@gmail.com (ORCID)

Authors:

Joseph Larmarange (ORCID)

Michael Curry (ORCID)

Jessica Lavery (ORCID)

Karissa Whiting (ORCID)

Emily C. Zabor (ORCID)

Other contributors:

Xing Bai [contributor]

Esther Drill (ORCID) [contributor]

Jessica Flynn (ORCID) [contributor]

Margie Hannum (ORCID) [contributor]

Stephanie Lobaugh [contributor]

Shannon Pileggi (ORCID) [contributor]

Amy Tin (ORCID) [contributor]

Gustavo Zapata Wainberg (ORCID) [contributor]