Report statistics from survfit tables inline

Extracts and returns statistics from a tbl_survfit object for inline reporting in an R markdown document. Detailed examples in the inline_text vignette

# S3 method for tbl_survfit
inline_text(
  x,
  variable = NULL,
  level = NULL,
  pattern = NULL,
  time = NULL,
  prob = NULL,
  column = NULL,
  estimate_fun = x$inputs$estimate_fun,
  pvalue_fun = NULL,
  ...
)

Arguments

x

Object created from tbl_survfit

variable

Variable name of statistic to present.

level

Level of the variable to display for categorical variables. Can also specify the 'Unknown' row. Default is NULL

pattern

String indicating the statistics to return.

time

time for which to return survival probabilities.

prob

probability with values in (0,1)

column

column to print from x$table_body. Columns may be selected with time= or prob= as well.

estimate_fun

Function to round and format estimate and confidence limits. Default is the same function used in tbl_survfit()

pvalue_fun

Function to round and format p-values. Default is style_pvalue. The function must have a numeric vector input (the numeric, exact p-value), and return a string that is the rounded/formatted p-value (e.g. pvalue_fun = function(x) style_pvalue(x, digits = 2) or equivalently, purrr::partial(style_pvalue, digits = 2)).

...

Not used

Value

A string reporting results from a gtsummary table

See also

Author

Daniel D. Sjoberg

Examples

library(survival)
# fit survfit
fit1 <- survfit(Surv(ttdeath, death) ~ trt, trial)
fit2 <- survfit(Surv(ttdeath, death) ~ 1, trial)

# sumarize survfit objects
tbl1 <-
  tbl_survfit(
    fit1,
    times = c(12, 24),
    label = ~"Treatment",
    label_header = "**{time} Month**"
  ) %>%
  add_p()

tbl2 <-
  tbl_survfit(
    fit2,
    probs = 0.5,
    label_header = "**Median Survival**"
  )

# report results inline
inline_text(tbl1, time = 24, level = "Drug B")
#> [1] "41% (33%, 52%)"
inline_text(tbl1, column = p.value)
#> [1] "p=0.2"
inline_text(tbl2, prob = 0.5)
#> [1] "22 (21, —)"