ARD summary table

The tbl_ard_summary() function tables descriptive statistics for continuous, categorical, and dichotomous variables. The functions accepts an ARD object.

Usage

tbl_ard_summary(
  cards,
  by = NULL,
  statistic = list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~
    "{n} ({p}%)"),
  type = NULL,
  missing = c("ifany", "no", "always"),
  missing_text = "Unknown",
  missing_stat = "{N_miss}",
  include = everything()
)

Arguments

cards

(card)
An ARD object of class "card" typically created with cards::ard_*() functions.

by

(tidy-select)
A single column from data. Summary statistics will be stratified by this variable. Default is NULL

statistic

(formula-list-selector)
Used to specify the summary statistics for each variable. Each of the statistics must be present in card as no new statistics are calculated in this function. The default is list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}%)").

type

(formula-list-selector)
Specifies the summary type. Accepted value are c("continuous", "continuous2", "categorical", "dichotomous"). Continuous summaries may be assigned c("continuous", "continuous2"), while categorical and dichotomous cannot be modified.

missing, missing_text, missing_stat

Arguments dictating how and if missing values are presented:

missing: must be one of c("ifany", "no", "always")

missing_text: string indicating text shown on missing row. Default is "Unknown"

missing_stat: statistic to show on missing row. Default is "{N_miss}". Possible values are N_miss, N_obs, N_nonmiss, p_miss, p_nonmiss

include

(tidy-select)
Variables to include in the summary table. Default is everything()

Value

a gtsummary table of class "tbl_ard_summary"

Examples

library(cards)

ard_stack(
  data = ADSL,
  ard_categorical(variables = "AGEGR1"),
  ard_continuous(variables = "AGE"),
  .attributes = TRUE,
  .missing = TRUE
) |>
  tbl_ard_summary()

  
  

Characteristic
      	N = 2541
    

Pooled Age Group 1

    65-80
	144 (56.7%)

    <65
	33 (13.0%)

    >80
	77 (30.3%)

Age
	77.0 (70.0, 81.0)

1 n (%); Median (25th Percentile, 75th Percentile)
    

ard_stack(
  data = ADSL,
  .by = ARM,
  ard_categorical(variables = "AGEGR1"),
  ard_continuous(variables = "AGE"),
  .attributes = TRUE,
  .missing = TRUE
) |>
  tbl_ard_summary(by = ARM)

  
  

Characteristic
      	Placebo

N = 861
      	Xanomeline High Dose

N = 841
      	Xanomeline Low Dose

N = 841
    

Pooled Age Group 1

    65-80
	42 (48.8%)
	55 (65.5%)
	47 (56.0%)

    <65
	14 (16.3%)
	11 (13.1%)
	8 (9.5%)

    >80
	30 (34.9%)
	18 (21.4%)
	29 (34.5%)

Age
	76.0 (69.0, 82.0)
	76.0 (70.5, 80.0)
	77.5 (71.0, 82.0)

1 n (%); Median (25th Percentile, 75th Percentile)