Methods for tbl_regression

Most regression models are handled by tbl_regression.default(), which uses broom::tidy() to perform initial tidying of results. There are, however, some model types that have modified default printing behavior. Those methods are listed below.

# S3 method for model_fit
tbl_regression(x, ...)

# S3 method for workflow
tbl_regression(x, ...)

# S3 method for survreg
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom::tidy(x, ...) %>% dplyr::filter(.data$term !=
    "Log(scale)"),
  ...
)

# S3 method for mira
tbl_regression(x, tidy_fun = pool_and_tidy_mice, ...)

# S3 method for mipo
tbl_regression(x, ...)

# S3 method for lmerMod
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

# S3 method for glmerMod
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

# S3 method for glmmTMB
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

# S3 method for glmmadmb
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

# S3 method for stanreg
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

# S3 method for brmsfit
tbl_regression(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

# S3 method for gam
tbl_regression(x, tidy_fun = tidy_gam, ...)

# S3 method for tidycrr
tbl_regression(x, tidy_fun = tidycmprsk::tidy, ...)

# S3 method for crr
tbl_regression(x, ...)

# S3 method for multinom
tbl_regression(x, ...)

Arguments

x

Regression model object

...

arguments passed to tbl_regression.default()

tidy_fun

Option to specify a particular tidier function for the model. Default is to use broom::tidy(), but if an error occurs then tidying of the model is attempted with parameters::model_parameters(), if installed.

Methods

The default method for tbl_regression() model summary uses broom::tidy(x) to perform the initial tidying of the model object. There are, however, a few models that use modifications.

"parsnip/workflows": If the model was prepared using parsnip/workflows, the original model fit is extracted and the original x= argument is replaced with the model fit. This will typically go unnoticed; however,if you've provided a custom tidier in tidy_fun= the tidier will be applied to the model fit object and not the parsnip/workflows object.

"survreg": The scale parameter is removed, broom::tidy(x) %>% dplyr::filter(term != "Log(scale)")

"multinom": This multinomial outcome is complex, with one line per covariate per outcome (less the reference group)

"gam": Uses the internal tidier tidy_gam() to print both parametric and smooth terms.

"tidycrr": Uses the tidier tidycmprsk::tidy() to print the model terms.

"lmerMod", "glmerMod", "glmmTMB", "glmmadmb", "stanreg", "brmsfit": These mixed effects models use broom.mixed::tidy(x, effects = "fixed"). Specify tidy_fun = broom.mixed::tidy to print the random components.